refactor(canvas): split Canvas/index.tsx into focused hooks
Canvas/index.tsx dropped from 1368 to 426 lines (69% reduction) by
extracting concerns into nine new hooks plus a fullscreen-chip
subcomponent. The root component now matches the task spec's "hooks
binding + prop passing" shape — every feature surface (focus,
selection, context menu, node actions, edge handlers, render order,
palette commands, mouse plumbing, sync effects, fullscreen chip)
lives in its own file with a single responsibility.
